services/reporting-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.35
aiomysql==0.2.0
cryptography>=42.0.0
PyMySQL==1.1.0
pydantic-settings==2.5.2
influxdb-client==1.40.0
minio==7.2.0
reportlab==4.2.5
apscheduler==3.10.4
alembic==1.13.3
python-dotenv==1.0.1
